Move Player out of multiworld.py
